Ch05
Added a method to deal with named inventory items (which we need when using the store inventory to differentiate ownership)

Fixed bug where if you sold items and had enough money to buy an item that was already in your inventory (waiting to be purchased) the buy button would not be enabled (no drop event)

Fixed bug with player items placed in store inventory, and then closing the store window. The issue is that the player items would not know the inded value in the players inventory. Created method to check to first available spots when saving.