Ch10
Bug Fix: When running the game in a maximized window, the monsters in the battleUI would move to a lower left position when attacked. The fix was to use the resized cell position in which the monster image is located, and reset the image position so that the shake camera uses the correct values.